rename cleaning processors to a more commonly known name
